+++ nym-vpn-app/src/contexts/main/provider.tsx
@@ -35,6 +35,7 @@ function MainStateProvider({ children, init }: Props) {
     enableLewesProtocol: init.enableLewesProtocol,
     mixnetTrafficConfig: init.mixnetTrafficConfig,
     mixnetTrafficDefaults: init.mixnetTrafficDefaults,
+    splitTunnel: init.splitTunnel,
   });
 
   const { push } = useInAppNotify();