+++ nym-vpn-app/CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Lewes protocol settings toggle
+- Split tunneling for Windows
 
 ## [1.27.0] - 2026-03-31